+++ src/ByteDev.Testing/Providers/ISettingsProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace ByteDev.Testing.Providers
+{
+    public interface ISettingsProvider
+    {
+        TTestSettings GetSettings<TTestSettings>() where TTestSettings : class, new();
+    }
+}