+++ tests/ByteDev.Testing.UnitTests/TestSettingsTests.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Linq;
-using System.Reflection;
-using ByteDev.Collections;
+﻿using ByteDev.Testing.Providers;
+using NSubstitute;
 using NUnit.Framework;
 
 namespace ByteDev.Testing.UnitTests
@@ -10,53 +8,53 @@ namespace ByteDev.Testing.UnitTests
     public class TestSettingsTests
     {
         [TestFixture]
-        public class Constructor : TestSettingsTests
-        {
-            [Test]
-            public void WhenNoParam_ThenSetDefaults()
-            {
-                var sut = new TestSettings();
-
-                Assert.That(sut.FilePaths, Is.Empty);
-            }
-
-            [Test]
-            public void WhenAssemblyIsNull_ThenThrowException()
-            {
-                Assert.Throws<ArgumentNullException>(() => _ = new TestSettings(null));
-            }
-
-            [Test]
-            public void WhenInit_ThenSetsFilePaths()
-            {
-                string userName = Environment.UserName;
-
-                var sut = new TestSettings(Assembly.GetAssembly(typeof(TestSettingsTests)));
-
-                var result = sut.FilePaths.ToList();
-
-                Assert.That(result.Count, Is.EqualTo(5));
-                Assert.That(result.First(), Is.EqualTo(@"C:\Temp\ByteDev.Testing.UnitTests.settings.json"));
-                Assert.That(result.Second(), Is.EqualTo(@"C:\Dev\ByteDev.Testing.UnitTests.settings.json"));
-                Assert.That(result.Third(), Is.EqualTo(@"Z:\Dev\ByteDev.Testing.UnitTests.settings.json"));
-                Assert.That(result.Fourth(), Is.EqualTo(@"C:\Users\" + userName + @"\ByteDev.Testing.UnitTests.settings.json"));
-                Assert.That(result.Fifth(), Is.EqualTo(@"C:\Users\" + userName + @"\Documents\ByteDev.Testing.UnitTests.settings.json"));
-            }
-        }
-
-        [TestFixture]
         public class GetSettings : TestSettingsTests
         {
-            [Test]
-            public void WhenNoKvConfig_AndFilePathsIsEmpty_ThenThrowException()
-            {
-                var sut = new TestSettings(Assembly.GetAssembly(typeof(TestSettingsTests)));
-                
-                sut.FilePaths.Clear();
-                sut.KeyVaultConfig.KeyVaultUri = null;
+            private TestSettings _sut;
 
-                var ex = Assert.Throws<TestingException>(() => _ = sut.GetSettings<object>());
-                Assert.That(ex.Message, Is.EqualTo("Could not create new settings instance."));
+            [SetUp]
+            public void SetUp()
+            {
+                _sut = new TestSettings();
+            }
+
+            [Test]
+            public void WhenNoProvidersAdded_ThenThrowException()
+            {
+                var ex = Assert.Throws<TestingException>(() => _ = _sut.GetSettings<object>());
+                Assert.That(ex.Message, Is.EqualTo("No settings providers added. Add at least one."));
+            }
+
+            [Test]
+            public void WhenOneProviderAdded_AndReturnsNull_ThenThrowException()
+            {
+                var provider = Substitute.For<ISettingsProvider>();
+
+                provider.GetSettings<object>().Returns(null);
+
+                _sut.AddProvider(provider);
+
+                var ex = Assert.Throws<TestingException>(() => _ = _sut.GetSettings<object>());
+                Assert.That(ex.Message, Is.EqualTo("Could not create new test settings instance."));
+            }
+
+            [Test]
+            public void WhenTwoProvidersAdded_AndSecondReturnsInstance_ThenReturnInstance()
+            {
+                var settings = new object();
+                
+                var provider1 = Substitute.For<ISettingsProvider>();
+                var provider2 = Substitute.For<ISettingsProvider>();
+
+                provider1.GetSettings<object>().Returns(null);
+                provider2.GetSettings<object>().Returns(settings);
+
+                _sut.AddProvider(provider1);
+                _sut.AddProvider(provider2);
+
+                var result = _sut.GetSettings<object>();
+
+                Assert.That(result, Is.SameAs(settings));
             }
         }
     }